Fixes and minor tweaks for the pull request
